Move the configuration for TLS to its own file.
